ParamInfo: added alias property
