+++ getdist/yaml_format_tools.py
@@ -18,6 +18,7 @@ _p_label = "latex"
 _p_dist = "dist"
 _p_value = "value"
 _p_derived = "derived"
+_p_alias = "alias"
 _separator = "__"
 _minuslogprior = "minuslogprior"
 _prior_1d_name = "0"
